// packages/angular/ac-angular/src/lib/ac-filters/enums/ac-filter-operators.enum.ts
export enum AcEnumFilterComparisonOperator{
  between = "between",
  equals = "equals",
  greaterThan = "greater_than",
  greaterThanOrEquals = "greater_than_or_equals",
  in = "in",
  isEmpty = "is_empty",
  isNull = "is_null",
  isNotNull = "is_not_null",
  lessThan = "less_than",
  lessThanOrEquals = "less_than_or_equals",
  like = "like",
  notBetween = "not_between",
  notEqual = "not_equal",
  notIn = "not_in",
}

export enum AcEnumFilterLogicalOperator{
  and = "and",
  or = "or"
}